Test coverage extension for FormService and ImageService
