+++ kie-server-parent/kie-server-tests/kie-server-integ-tests-all/src/test/java/org/kie/server/integrationtests/jbpm/ImageServiceIntegrationTest.java
@@ -15,37 +15,22 @@
 
 package org.kie.server.integrationtests.jbpm;
 
-import java.util.HashMap;
-import java.util.Map;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
-
-import org.jboss.resteasy.client.ClientRequest;
-import org.jboss.resteasy.client.ClientResponse;
-import org.jboss.resteasy.client.ClientResponseFailure;
-import org.junit.Assert;
 import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.ClassRule;
 import org.junit.Test;
 import org.junit.rules.ExternalResource;
-import org.kie.server.api.marshalling.Marshaller;
-import org.kie.server.api.marshalling.MarshallerFactory;
 import org.kie.server.api.model.KieContainerResource;
 import org.kie.server.api.model.ReleaseId;
-import org.kie.server.api.model.type.JaxbLong;
-import org.kie.server.api.rest.RestURI;
 import org.kie.server.client.KieServicesClient;
 import org.kie.server.client.KieServicesException;
 import org.kie.server.client.ProcessServicesClient;
 import org.kie.server.client.UIServicesClient;
-import org.kie.server.integrationtests.config.TestConfig;
-import org.kie.server.integrationtests.shared.RestOnlyBaseIntegrationTest;
+import org.kie.server.integrationtests.shared.RestJmsSharedBaseIntegrationTest;
 
 import static org.junit.Assert.*;
-import static org.kie.server.api.rest.RestURI.*;
 
-public class ImageServiceIntegrationTest extends RestOnlyBaseIntegrationTest {
+public class ImageServiceIntegrationTest extends RestJmsSharedBaseIntegrationTest {
 
     private static ReleaseId releaseId = new ReleaseId("org.kie.server.testing", "definition-project",
             "1.0.0.Final");
@@ -69,7 +54,7 @@ public class ImageServiceIntegrationTest extends RestOnlyBaseIntegrationTest {
     }
 
     @Override
-    protected KieServicesClient createDefaultClient() {
+    protected KieServicesClient createDefaultClient() throws Exception {
         KieServicesClient servicesClient = super.createDefaultClient();
 
         uiServicesClient = servicesClient.getServicesClient(UIServicesClient.class);
@@ -82,158 +67,6 @@ public class ImageServiceIntegrationTest extends RestOnlyBaseIntegrationTest {
     public void cleanup() {
         cleanupSingletonSessionId();
 
-    }
-
-    @Test
-    public void testGetProcessImageTest() throws Exception {
-        KieContainerResource resource = new KieContainerResource(CONTAINER_ID, releaseId);
-        assertSuccess(client.createContainer(CONTAINER_ID, resource));
-
-        Map<String, Object> valuesMap = new HashMap<String, Object>();
-        valuesMap.put(RestURI.CONTAINER_ID, resource.getContainerId());
-        valuesMap.put(RestURI.PROCESS_ID, HIRING_PROCESS_ID);
-
-        ClientResponse<?> response = null;
-        try {
-
-            ClientRequest clientRequest = newRequest(build(TestConfig.getKieServerHttpUrl(), IMAGE_URI + "/" + PROCESS_IMG_GET_URI, valuesMap))
-                    .header("Content-Type", getMediaType().toString())
-                    .header("Accept", MediaType.APPLICATION_SVG_XML);
-            logger.info( "[GET] " + clientRequest.getUri());
-
-            response = clientRequest.get();
-            Assert.assertEquals(Response.Status.OK.getStatusCode(), response.getStatus());
-
-            String result = response.getEntity(String.class);
-            logger.debug("Image content is '{}'", result);
-            assertNotNull(result);
-            assertFalse(result.isEmpty());
-
-
-        } catch (Exception e) {
-            throw new ClientResponseFailure(e, response);
-        }  finally {
-            if (response != null) {
-                response.releaseConnection();
-            }
-        }
-
-    }
-
-    @Test
-    public void testGetProcessInstanceImageTest() throws Exception {
-        KieContainerResource resource = new KieContainerResource(CONTAINER_ID, releaseId);
-        assertSuccess(client.createContainer(CONTAINER_ID, resource));
-
-        Map<String, Object> valuesMap = new HashMap<String, Object>();
-        valuesMap.put(RestURI.CONTAINER_ID, resource.getContainerId());
-        valuesMap.put(RestURI.PROCESS_ID, HIRING_PROCESS_ID);
-
-        Marshaller marshaller = MarshallerFactory.getMarshaller(marshallingFormat, ClassLoader.getSystemClassLoader());
-
-        ClientResponse<?> response = null;
-        try {
-
-            Map<String, Object> params = new HashMap<String, Object>();
-            params.put("name", "john");
-
-            // start process instance
-            ClientRequest clientRequest = newRequest(build(TestConfig.getKieServerHttpUrl(), PROCESS_URI + "/" + START_PROCESS_POST_URI, valuesMap))
-                    .header("Content-Type", getMediaType().toString())
-                    .header("Accept", getMediaType().toString())
-                    .body(getMediaType(), marshaller.marshall(params));
-            logger.info("[POST] " + clientRequest.getUri());
-            response = clientRequest.post();
-            Assert.assertEquals(Response.Status.CREATED.getStatusCode(), response.getStatus());
-
-            Long result = response.getEntity(JaxbLong.class).unwrap();
-            assertNotNull(result);
-
-            valuesMap.put(RestURI.PROCESS_INST_ID, result);
-
-            clientRequest = newRequest(build(TestConfig.getKieServerHttpUrl(), IMAGE_URI + "/" + PROCESS_INST_IMG_GET_URI, valuesMap))
-                    .header("Content-Type", getMediaType().toString())
-                    .header("Accept", MediaType.APPLICATION_SVG_XML);
-            logger.info("[GET] " + clientRequest.getUri());
-
-            response = clientRequest.get();
-            Assert.assertEquals(Response.Status.OK.getStatusCode(), response.getStatus());
-
-            String image = response.getEntity(String.class);
-            logger.debug("Image content is '{}'", image);
-            assertNotNull(image);
-            assertFalse(image.isEmpty());
-
-            clientRequest = newRequest(build(TestConfig.getKieServerHttpUrl(), PROCESS_URI + "/" + ABORT_PROCESS_INST_DEL_URI, valuesMap)).header("Content-Type", getMediaType().toString());
-            logger.info("[DELETE] " + clientRequest.getUri());
-            response = clientRequest.delete();
-            Assert.assertEquals(Response.Status.NO_CONTENT.getStatusCode(), response.getStatus());
-
-        } catch (Exception e) {
-            throw new ClientResponseFailure(e, response);
-        } finally {
-            if (response != null) {
-                response.releaseConnection();
-            }
-        }
-    }
-
-    @Test
-    public void testGetProcessImageNotExistingTest() throws Exception {
-        KieContainerResource resource = new KieContainerResource(CONTAINER_ID, releaseId);
-        assertSuccess(client.createContainer(CONTAINER_ID, resource));
-
-        Map<String, Object> valuesMap = new HashMap<String, Object>();
-        valuesMap.put(RestURI.CONTAINER_ID, resource.getContainerId());
-        valuesMap.put(RestURI.PROCESS_ID, "not-existing");
-
-        ClientResponse<?> response = null;
-        try {
-
-            ClientRequest clientRequest = newRequest(build(TestConfig.getKieServerHttpUrl(), IMAGE_URI + "/" + PROCESS_IMG_GET_URI, valuesMap))
-                    .header("Content-Type", getMediaType().toString())
-                    .header("Accept", MediaType.APPLICATION_SVG_XML);
-            logger.info( "[GET] " + clientRequest.getUri());
-
-            response = clientRequest.get();
-            Assert.assertEquals(Response.Status.NOT_FOUND.getStatusCode(), response.getStatus());
-
-        } catch (Exception e) {
-            throw new ClientResponseFailure(e, response);
-        }  finally {
-            if (response != null) {
-                response.releaseConnection();
-            }
-        }
-    }
-
-    @Test
-    public void testGetProcessInstanceImageNotExistingTest() throws Exception {
-        KieContainerResource resource = new KieContainerResource(CONTAINER_ID, releaseId);
-        assertSuccess(client.createContainer(CONTAINER_ID, resource));
-
-        Map<String, Object> valuesMap = new HashMap<String, Object>();
-        valuesMap.put(RestURI.CONTAINER_ID, resource.getContainerId());
-        valuesMap.put(RestURI.PROCESS_INST_ID, 9999);
-
-        ClientResponse<?> response = null;
-        try {
-
-            ClientRequest clientRequest = newRequest(build(TestConfig.getKieServerHttpUrl(), IMAGE_URI + "/" + PROCESS_INST_IMG_GET_URI, valuesMap))
-                    .header("Content-Type", getMediaType().toString())
-                    .header("Accept", MediaType.APPLICATION_SVG_XML);
-            logger.info( "[GET] " + clientRequest.getUri());
-
-            response = clientRequest.get();
-            Assert.assertEquals(Response.Status.NOT_FOUND.getStatusCode(), response.getStatus());
-
-        } catch (Exception e) {
-            throw new ClientResponseFailure(e, response);
-        }  finally {
-            if (response != null) {
-                response.releaseConnection();
-            }
-        }
     }
 
     @Test
